Edit html containers, add flex properties
First added the content container to the header, close button, text and
the continue/cancel buttons. Expanded the header class to include the
cancel button so those would be vertically aligned.

Changed the display property to flex for the modal then started aligning
classes in that modal to get it looking like the desired outcome.