Finish alignment and style elements
To get everything aligned properly I made the body flex and changed
its alignment to column. Then I made the footer flex to get it all
horizontal and added margin-top:auto to stick it to the bottom.
Once the header and footer were in place I added margin: auto to the
content div to center it.

After alignment I styled the fonts, added padded and such to get the
page looking nice and like the desired outcome.